Fixed Client side: goroutine "Bug", simplification of toy client
- i understood the problem where the caller of the goroutine,
did not wait for the callee. Resulting in a "closed network conn"
error. note: there are two "flavours" of this problem. all
similar! ask me for more information :D

- i also removed debug code

- i also removed the usage of a channel in the toy client.
we dont need to be able to process the reply send from the server
by the client atm. i.e. client sends a request, server processes
and sends reply, client receives and processes reply

- we also send a dummy json atm

Signed-off-by: monfron <simon.fronhoefer@gmail.com>